chore(console): clean up
